Refactor fetching providers to its own common reducer
