+++ app/javascript/react/screens/App/Overview/OverviewConstants.js
@@ -20,9 +20,6 @@ export const HIDE_CONFIRM_MODAL = 'HIDE_CONFIRM_MODAL';
 export const ARCHIVE_TRANSFORMATION_PLAN = 'ARCHIVE_TRANSFORMATION_PLAN';
 export const DELETE_TRANSFORMATION_PLAN = 'DELETE_TRANSFORMATION_PLAN';
 
-export const FETCH_PROVIDERS = 'FETCH_PROVIDERS';
-export const FETCH_PROVIDERS_URL = '/api/providers?expand=resources&attributes=authentications';
-
 export const FETCH_V2V_TRANSFORMATION_MAPPINGS = 'FETCH_V2V_TRANSFORMATION_MAPPINGS';
 
 export const FETCH_V2V_TRANSFORMATION_PLANS = 'FETCH_V2V_TRANSFORMATION_PLANS';
@@ -45,11 +42,6 @@ export const OSP_TENANT = 'CloudTenant';
 export const OSP_VOLUME = 'CloudVolumeType';
 export const OSP_NETWORK = 'CloudNetwork';
 export const EMS_CLUSTER = 'EmsCluster';
-
-export const PROVIDERS = {
-  source: ['ManageIQ::Providers::Vmware::InfraManager'],
-  target: ['ManageIQ::Providers::Redhat::InfraManager', 'ManageIQ::Providers::Openstack::CloudManager']
-};
 
 export const FETCH_TRANSFORMATION_PLANS_URL =
   '/api/service_templates/?' +
