+++ app/javascript/react/screens/App/Overview/OverviewValidators.js
@@ -1,16 +1,11 @@
 import { validateSchema } from '../../../../data/schemaHelpers';
 import { plansSchema } from '../../../../data/models/plans';
-import { providersSchema } from '../../../../data/models/provider';
 import { requestsSchema } from '../../../../data/models/requests';
 import { mappingsSchema } from '../../../../data/models/transformationMappings';
 import { serviceTemplatePlaybookSchema } from '../../../../data/models/playbooks';
 
 export const validateOverviewPlans = plans => {
   validateSchema(plansSchema, plans);
-};
-
-export const validateOverviewProviders = providers => {
-  validateSchema(providersSchema, providers);
 };
 
 export const validateOverviewRequests = requests => {
