+++ app/javascript/react/screens/App/Overview/__test__/helpers.test.js
@@ -1,4 +1,4 @@
-import { attachTargetProvider, hasRsaKey } from '../helpers';
+import { attachTargetProvider } from '../helpers';
 import { TRANSFORMATION_MAPPING_ITEM_DESTINATION_TYPES } from '../../Mappings/screens/MappingWizard/MappingWizardConstants';
 
 describe('attachTargetProvider', () => {
@@ -27,21 +27,5 @@ describe('attachTargetProvider', () => {
     const result = attachTargetProvider(plan, [provider], clusters, 'rhevm');
 
     expect(result).toEqual({ ...plan, targetProvider: provider });
-  });
-});
-
-describe('hasRsaKey', () => {
-  test('returns false if the provider is missing the authentications attribute', () => {
-    const provider = { name: 'provider' };
-    const result = hasRsaKey(provider);
-
-    expect(result).toBe(false);
-  });
-
-  test('returns true if the provider has an authentication of type ssh_keypair', () => {
-    const provider = { authentications: [{ authtype: 'ssh_keypair' }] };
-    const result = hasRsaKey(provider);
-
-    expect(result).toBe(true);
   });
 });
