+++ app/javascript/react/screens/App/Overview/index.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 import Overview from './Overview';
 import * as OverviewActions from './OverviewActions';
 import * as NotificationActions from '../common/NotificationList/NotificationListActions';
+import * as ProvidersActions from '../../../../redux/common/providers/providersActions';
 import * as RouterActions from '../../../../redux/actions/routerActions';
 
 import reducer from './OverviewReducer';
@@ -17,18 +18,19 @@ import {
 export const reducers = { overview: reducer, form: {} };
 
 const mapStateToProps = (
-  { overview, overview: { transformationPlans, allRequestsWithTasks, planId, providers, cloudTenants } },
+  { overview, overview: { transformationPlans, allRequestsWithTasks, planId, cloudTenants }, providers },
   ownProps
 ) => {
   const { requestsProcessingCancellation, ...overviewRest } = overview;
 
   return {
     ...overviewRest,
+    ...providers,
     ...ownProps.data,
     notStartedTransformationPlans: notStartedTransformationPlansFilter(transformationPlans),
     activeTransformationPlans: attachTargetProviderToOspPlans(
       activeTransformationPlansFilter(transformationPlans, planId),
-      providers,
+      providers.providers,
       cloudTenants
     ),
     finishedTransformationPlans: finishedTransformationPlansFilter(transformationPlans),
@@ -43,6 +45,6 @@ const mergeProps = (stateProps, dispatchProps, ownProps) => Object.assign(stateP
 
 export default connect(
   mapStateToProps,
-  Object.assign(OverviewActions, NotificationActions, RouterActions),
+  Object.assign(OverviewActions, NotificationActions, ProvidersActions, RouterActions),
   mergeProps
 )(Overview);
