+++ app/javascript/react/screens/App/Settings/SettingsConstants.js
@@ -5,5 +5,4 @@ export const FETCH_V2V_CONVERSION_HOSTS = 'FETCH_V2V_CONVERSION_HOSTS';
 export const SHOW_V2V_CONVERSION_HOST_WIZARD = 'SHOW_V2V_CONVERSION_HOST_WIZARD';
 export const HIDE_V2V_CONVERSION_HOST_WIZARD = 'HIDE_V2V_CONVERSION_HOST_WIZARD';
 export const V2V_CONVERSION_HOST_WIZARD_EXITED = 'V2V_CONVERSION_HOST_WIZARD_EXITED';
-export const FETCH_V2V_PROVIDERS = 'FETCH_V2V_PROVIDERS';
 export const FETCH_V2V_TARGET_COMPUTE_RESOURCES = 'FETCH_V2V_TARGET_COMPUTE_RESOURCES';
