+++ app/javascript/redux/common/providers/providers.fixtures.js
