+++ app/javascript/redux/common/providers/providersReducer.js
@@ -1,0 +1,48 @@
+import Immutable from 'seamless-immutable';
+
+import { FETCH_V2V_PROVIDERS } from './providersConstants';
+import { validateProviders } from './providersValidators';
+import { hasRsaKey, sufficientProviders } from './providersHelpers';
+
+export const initialState = Immutable({
+  isFetchingProviders: false,
+  isRejectedProviders: false,
+  errorFetchingProviders: null,
+  providers: [],
+  hasSufficientProviders: false
+});
+
+export default (state = initialState, action) => {
+  switch (action.type) {
+    case `${FETCH_V2V_PROVIDERS}_PENDING`:
+      return state
+        .set('isFetchingProviders', true)
+        .set('isRejectedProviders', false)
+        .set('errorFetchingProviders', null);
+    case `${FETCH_V2V_PROVIDERS}_FULFILLED`:
+      return (() => {
+        const insufficient = state
+          .set('hasSufficientProviders', false)
+          .set('isFetchingProviders', false)
+          .set('isRejectedProviders', false)
+          .set('errorFetchingProviders', null);
+        if (!action.payload.data || !action.payload.data.resources) {
+          return insufficient;
+        }
+        validateProviders(action.payload.data.resources);
+        return insufficient
+          .set('hasSufficientProviders', sufficientProviders(action.payload.data.resources))
+          .set(
+            'providers',
+            action.payload.data.resources.map(provider => ({ ...provider, hasRsaKey: hasRsaKey(provider) }))
+          );
+      })();
+    case `${FETCH_V2V_PROVIDERS}_REJECTED`:
+      return state
+        .set('isFetchingProviders', false)
+        .set('isRejectedProviders', true)
+        .set('errorFetchingProviders', action.payload);
+    default:
+      return state;
+  }
+};
