+++ app/javascript/redux/common/providers/providersValidators.js
@@ -1,0 +1,6 @@
+import { validateSchema } from '../../../data/schemaHelpers';
+import { providersSchema } from '../../../data/models/provider';
+
+export const validateProviders = providers => {
+  validateSchema(providersSchema, providers);
+};
